Moved scripts from db_access folder to db_access.py #126
