--- docs/source/modules/db_access.rst
@@ -1,42 +1,0 @@
-db_access subpackage
-====================
-
-biocyc module
--------------
-
-.. automodule:: refinegems.curation.db_access.biocyc
-   :members:
-   :undoc-members:
-   :show-inheritance:
-   :private-members: 
-   :special-members:
-
-db module
----------
-
-.. automodule:: refinegems.curation.db_access.db
-   :members:
-   :undoc-members:
-   :show-inheritance:
-   :private-members: 
-   :special-members:
-
-kegg module
------------
-
-.. automodule:: refinegems.curation.db_access.kegg
-   :members:
-   :undoc-members:
-   :show-inheritance:
-   :private-members: 
-   :special-members:
-
-modelseed module
-----------------
-
-.. automodule:: refinegems.curation.db_access.modelseed
-   :members:
-   :undoc-members:
-   :show-inheritance:
-   :private-members: 
-   :special-members:
