--- src/refinegems/curation/db_access/__init__.py
@@ -1,6 +1,0 @@
-__all__ = ['biocyc','db','kegg','modelseed']
-
-from . import biocyc
-from . import db
-from . import kegg
-from . import modelseed
